ENH: post-review restructure
