+++ pysatMissionPlanning/__init__.py
@@ -2,9 +2,11 @@ from __future__ import print_function
 from __future__ import absolute_import
 
 import os
-from pysatMissionPlanning import instruments, methods
+from pysatMissionPlanning import instruments
+from pysatMissionPlanning import methods
+from pysatMissionPlanning import plot
 
-__all__ = ['instruments', 'methods']
+__all__ = ['instruments', 'methods', 'plot']
 
 # set version
 here = os.path.abspath(os.path.dirname(__file__))
