+++ pysatMissionPlanning/instruments/pysat_ephem.py
@@ -15,7 +15,7 @@ import functools
 import numpy as np
 import pandas as pds
 import pysat
-import pysatMissionPlanning.methods.simulators as sim
+import pysatMissionPlanning.instruments.supporting_methods as meth
 
 # pysat required parameters
 platform = 'pysat'
@@ -185,8 +185,8 @@ def load(fnames, tag=None, sat_id=None, obs_long=0., obs_lat=0., obs_alt=0.,
     return data, meta.copy()
 
 
-list_files = functools.partial(sim.list_files)
-download = functools.partial(sim.download)
+list_files = functools.partial(meth.list_files)
+download = functools.partial(meth.download)
 
 # create metadata corresponding to variables in load routine just above
 # made once here rather than regenerate every load call
