+++ pysatMissionPlanning/instruments/supporting_methods.py
