+++ pysatMissionPlanning/methods/__init__.py
@@ -5,9 +5,7 @@ the methods to interface with numerous empirical model packages
 
 from pysatMissionPlanning.methods import aacgmv2
 from pysatMissionPlanning.methods import apexpy
-from pysatMissionPlanning.methods import plot
 from pysatMissionPlanning.methods import pyglow
-from pysatMissionPlanning.methods import simulators
 from pysatMissionPlanning.methods import spacecraft
 
-__all__ = ['aacgmv2', 'apexpy', 'plot', 'pyglow', 'simulators', 'spacecraft']
+__all__ = ['aacgmv2', 'apexpy', 'pyglow', 'spacecraft']
